changed the manner of changing states and got rid of the enum for the states
The sates were previously in an enum but you can't increment an enum so the sates now are all defines. The next states are set but nothing is done at the moment of setting it. There is another fsm that will change the interupt trigger and the data line. There was a lot of repetion so this makes less repetitive